# controllers/books/book.dto.go
package controllers

type CreateBookInput struct {
	Title  string `json:"title" binding:"required"`
	Author string `json:"author" binding:"required"`
}

type UpdateBookInput struct {
	Title  string `json:"title"`
	Author string `json:"author"`
}

type BookOutput struct {
	ID     uint
	Title  string
	Author string
}